pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools-scm>=8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asguardian"
dynamic = ["version"]
description = "Asguardian - Universal Development Tools Suite"
readme = "README.md"
license = {text = "Polyform Noncommercial 1.0.0"}
requires-python = ">=3.11"
authors = [
    {name = "Asgard Contributors"}
]
keywords = [
    "code-quality",
    "static-analysis",
    "security",
    "testing",
    "api-validation",
    "schema-validation",
    "monitoring",
    "developer-tools",
    "linting",
    "architecture",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
asguardian = "Asgard.cli:main"
heimdall = "Asgard.Heimdall.cli:main"
freya = "Asgard.Freya.cli:main"
forseti = "Asgard.Forseti.cli:main"
verdandi = "Asgard.Verdandi.cli:main"
volundr = "Asgard.Volundr.cli:main"
asguardian-mcp = "Asgard.MCP.server:main"
asguardian-dashboard = "Asgard.Dashboard.server:main"

[project.urls]
Homepage = "https://github.com/JakeDruett/asguardian"
Documentation = "https://github.com/JakeDruett/asguardian#readme"
Repository = "https://github.com/JakeDruett/asguardian"

[tool.setuptools.packages.find]
exclude = ["Asgard_Test", "Asgard_Test.*"]

[tool.pytest.ini_options]
testpaths = ["Asgard_Test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "unit: unit tests",
    "integration: integration tests",
    "volundr: Volundr package tests",
    "heimdall: Heimdall package tests",
    "freya: Freya package tests",
    "verdandi: Verdandi package tests",
    "forseti: Forseti package tests",
    "backend_init: BackendInit package tests",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM108", # use ternary operator (can reduce readability)
]

[tool.ruff.lint.isort]
known-first-party = ["Asgard"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["Asgard"]
omit = [
    "Asgard_Test/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
show_missing = true
precision = 2
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "raise NotImplementedError",
]

[tool.coverage.html]
directory = "htmlcov"

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"
